+3 Ar 3-21-2018 Optional admin prefix Variable Supported
The feature Supported in react-scripts@0.5.0 or higher (current is
react-scripts@1.0.17). How it works is the default is empty string if
you don't need this modification for the admin panel, so for deployment
which do not need a prefix for js and css resources in admin panel,
simple leave this alone and don't make use of this option. However, now
supported is the following option: to put a optional admin prefix for js
and css resources duch as "/admin" for certain nginx configurations, how
you do it is at the project root, create a file called ".env" Within the
file called .env what you should do is place this line:
REACT_APP_OPTIONAL_ADMIN_PREFIX=/admin and then after that save the .env
file - after that you should restart the react app server. Now the js
and css resources will come from /admin if you need it.